Change the value type of the minimum and maximum attribute
Accordingly the html documentation, the value type of the html attribute 'min' or 'max' depends on the html element. If the element is a 'input' the value could be a number or a date. If the element is a 'meter' or 'progress' the value should be just a floating-point number.