Attributes improvements (#76)
* Remove the resolved-file.

* Update the gitignore-file

* Add the missing InputElement protocol to the GlobalElement alias

* Remove the case Self from the enum Target to resolve the warning introduced in Swift 5.6

* Revise comments

* Introduce a new protocol AnyLayout to reduce repeating code in this package and also in the vapor-provider package.

* Relocate the converter stencils

* Relocate the CustomAttribute in the AnyAttribute protocol, so it is not necessary to have them both

* Add a mutate-method for the SelectedAttribute if Self is a EmptyNode

* Change the value type of the accesskey attribute from string to character.

* Change the value type of the translate attribute

Accordingly the html documentation, the html attribute 'translate' requires a 'yes' or 'no' as value. Change the value type of the attribute 'translate' from string to the htmlkit type 'decision'. It prevents the user from entering a wrong value or something unrelated.

* Change the value type of the kind attribute

Accordingly the html documentation, the html attribute 'kind' requires specific values. Change the value type of the attribute 'kind' from string to the htmlkit type 'kinds'. It prevents the user from entering a wrong value or something unrelated.

* Change the value type of the low attribute

Accordingly the html documentation, the html attribute 'low' requires a floating-point number as value. Change the value type of the attribute 'low' from integer to float.

* Change the value type of the high attribute

Accordingly the html documentation, the html attribute 'high' requires a floating-point number as value. Change the value type of the attribute 'high' from integer to float.

* Change the value type of the minimum and maximum attribute

Accordingly the html documentation, the value type of the html attribute 'min' or 'max' depends on the html element. If the element is a 'input' the value could be a number or a date. If the element is a 'meter' or 'progress' the value should be just a floating-point number.

* Change the value type of the maximum/minimum length attribute

Accordingly the html documentation the html attributes 'maxlength' and 'minlength' require an integer as a value. Change the value type from string to integer.

* Change the value type of the tabulator attribute

Accordingly the html documentation, the html attribute 'tabindex' requires an integer as value. Change the value type from string to integer.

* Change the value type of the preload attribute and add the missing methods

* Add tests for the attribute rendering