Add a plugin for the conversion (#97)
* Separate the converter from htmlkit and declare it as an independent target

* Add the converter plugin

* Refactor the converter

Separate the parsing method from the converter by creating a new class and also use the string conversion as the primary method.

* Add the convert command

* Update the converter plugin

* Refactor the converter

Change the access of the method for the string conversion to private . Also unify the conversion methods for files and directories.

* Update the convert command

* Update the converter plugin

* Add a check for the output option

Debug and file should be the only allowed options. The check should make sure of it.