Refactor code (#101)
* Remove deprecations file

* Remove thumbnail

* Rename the documentation catalog of htmlkit

* Refactor the formula class of the renderer

* Extract the cache handling in its own class

The class should handle the caching later and everything that comes with.

* Refactor the renderer class

The functions add(_ :) and render(_ :) can be simplified. Therefore the view requirements have to be removed.

* Restore renderer methods

Brings back the briefly removed methods and mark them as deprecated.

* Fix test failing

* Extract the localization in its own class

The class should handle the localization later and everything that comes with it.

* Remove the modifiable protocol

* Move the draw handler for the path element way up

The draw attribute on the group element is incorrect. It should be removed.

* Change the initializer of the document element

* Add computed properties to the node protocols

It reduces the code redundancy, plus each node is one single ingredient in the renderer formula.

* Remove the string builder

The converter doesn't use it anymore.

* Rework the context manager

* Remove the context variable file

The html context class is slightly the same. While the html context is more used, the context variable file can be removed.

* Add tests for the statements

* Remove the template value mapping file

* Refactor the html context class

* Refactor the template value enum

* Change the access level of the conditions

The access level for the comparison operators stays public.

* Add a namespace for the several enumerations of attribute values

* Adjust the syntax of some attribute handlers

* Rearrange and rename files

* Refactor the statements

* Change the internal comments

* Add tests for the conditions

* Refactor the conditions